+++ src/meltano/api/security/__init__.py
@@ -2,14 +2,19 @@ import urllib.parse
 from datetime import timedelta
 from functools import wraps
 from flask import current_app, request, redirect, jsonify, make_response
-from flask_login import current_user
+from flask_login import current_user, user_logged_in
 from flask_security import Security, login_required
 from flask_security.utils import login_user
 from flask_principal import identity_loaded, Identity
 
 from .identity import users, FreeUser, create_dev_user
 from .forms import MeltanoLoginForm, MeltanoRegisterFrom, MeltanoConfirmRegisterForm
-from .auth import unauthorized_callback, _identity_loaded_hook, api_auth_required
+from .auth import (
+    unauthorized_callback,
+    _user_logged_in_hook,
+    _identity_loaded_hook,
+    api_auth_required,
+)
 
 
 # normally one would setup the extension accordingly, but it
@@ -31,4 +36,5 @@ def setup_security(app, project):
 
     security.init_app(app, users, **options)
     security.unauthorized_handler(unauthorized_callback)
+    user_logged_in.connect_via(app)(_user_logged_in_hook)
     identity_loaded.connect_via(app)(_identity_loaded_hook)